Added LFA of a Composite
This project was completed for a course on polymer simulation. It is based on previous work that embedded strands of a thermally conductive filament into an epoxy matrix to create samples for measuring the thermal diffusivity. No exisiting models could accurately capture the behavior of this complex system, so I developed a numerical technique based on a spectrum of conductive path lengths. This article was not submitted for publication due to a lack of experimental data and the development of better techniques for measuring these materials.